Updated RPATH logic
This applies to macOS
We expected that RPATH=OFF would have result in absolute path in library name.
Unfortunately this is not the case, as it seems that CMAKE_SKIP_RPATH controls
the RPATH in the executable, while MACOSX_RPATH controls the library.
INSTALL_NAME_DIR can be used to specify the absolute path of the library, but there are
two (unexpected) cases in which this do not apply.
This patch tries to solve this issue by (forcely) removing those two corner cases.

@drdanz